requirement.txt
asgiref==3.6.0
attrs==23.1.0
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.1.0
cryptography==40.0.2
Deprecated==1.2.13
Django==4.2.1
django-environ==0.10.0
django-oauth-toolkit==2.2.0
django-rest-framework-multi-slug-field==0.2.0
djangorestframework==3.14.0
drf-spectacular==0.26.2
idna==3.4
inflection==0.5.1
install==1.3.5
jsonschema==4.17.3
jwcrypto==1.4.2
oauthlib==3.2.2
psycopg2==2.9.6
pycparser==2.21
pyrsistent==0.19.3
pytz==2023.3
PyYAML==6.0
requests==2.30.0
sqlparse==0.4.4
tzdata==2023.3
uritemplate==4.1.1
urllib3==2.0.2
wrapt==1.15.0
